Move models to General directory
