+++ TMDbLib/Objects/General/WatchProviderItem.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace TMDbLib.Objects.Movies
+namespace TMDbLib.Objects.General
 {
     public class WatchProviderItem
     {
